+++ src/pages/Main.tsx
@@ -10,7 +10,12 @@ import {
 } from "react";
 import { useRecoilState } from "recoil";
 import createScrollSnap from "scroll-snap";
-import { MainContent, mainContent, mainScrollRef } from "../atoms/mainContent";
+import {
+  isMainsSrollUp,
+  MainContent,
+  mainContent,
+  mainScrollRef,
+} from "../atoms/mainContent";
 import {
   AnimatePresence,
   motion as m,
@@ -24,7 +29,7 @@ import main3 from "../image/main3.png";
 import main4 from "../image/main4.png";
 import main5 from "../image/main5.png";
 import main6 from "../image/main6.png";
-import { cls } from "../utils";
+import { cls, cme_scrollToY } from "../utils";
 import Portal from "../components/Portal";
 import { isScrolled } from "../atoms/isScrolled";
 import { IsScrolled } from "../types";
@@ -58,11 +63,7 @@ const ScrollTop = () => {
   const [contentNo] = useRecoilState<MainContent>(mainContent);
   const [getMainScrollRef] = useRecoilState(mainScrollRef);
   const handleClick = () => {
-    getMainScrollRef?.scrollTo({
-      left: 0,
-      top: 0,
-      behavior: "smooth",
-    });
+    cme_scrollToY(0, 1200, getMainScrollRef);
   };
   return (
     <Portal>
@@ -469,28 +470,30 @@ const EnterButton = ({ scrollYProgress, snapContainer }) => {
     isGoingUpOrStoped: true,
     isIntroing: true,
   });
+  const [getIsMainsSrollUp, setIsMainsSrollUp] =
+    useRecoilState<boolean>(isMainsSrollUp);
   const { innerHeight: screenHeight } = window;
-  function cme_scrollToY(y, duration = 0, element = document.scrollingElement) {
-    // cancel if already on target position
-    if (element.scrollTop === y) return;
+  // function cme_scrollToY(y, duration = 0, element = document.scrollingElement) {
+  //   // cancel if already on target position
+  //   if (element.scrollTop === y) return;
 
-    const cosParameter = (element.scrollTop - y) / 2;
-    let scrollCount = 0,
-      oldTimestamp = null;
+  //   const cosParameter = (element.scrollTop - y) / 2;
+  //   let scrollCount = 0,
+  //     oldTimestamp = null;
 
-    function step(newTimestamp) {
-      if (oldTimestamp !== null) {
-        // if duration is 0 scrollCount will be Infinity
-        scrollCount += (Math.PI * (newTimestamp - oldTimestamp)) / duration;
-        if (scrollCount >= Math.PI) return (element.scrollTop = y);
-        element.scrollTop =
-          cosParameter + y + cosParameter * Math.cos(scrollCount);
-      }
-      oldTimestamp = newTimestamp;
-      window.requestAnimationFrame(step);
-    }
-    window.requestAnimationFrame(step);
-  }
+  //   function step(newTimestamp) {
+  //     if (oldTimestamp !== null) {
+  //       // if duration is 0 scrollCount will be Infinity
+  //       scrollCount += (Math.PI * (newTimestamp - oldTimestamp)) / duration;
+  //       if (scrollCount >= Math.PI) return (element.scrollTop = y);
+  //       element.scrollTop =
+  //         cosParameter + y + cosParameter * Math.cos(scrollCount);
+  //     }
+  //     oldTimestamp = newTimestamp;
+  //     window.requestAnimationFrame(step);
+  //   }
+  //   window.requestAnimationFrame(step);
+  // }
   const handleClick = () => {
     const duration = 8200 * (1 - (16 * scrollYProgress.get()) / 10); // 어디서 눌러도 같은 속도로
     cme_scrollToY(screenHeight * 10, duration, snapContainer.current);
